pyproject.toml
[project]
name = "bank-api"
version = "0.1.0"
description = ""
authors = [
    {name = ""}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "pydantic (>=2.12.5,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
